chore(tooltip): replace current tooltip with v2
