--- packages/primitives/tooltip-v2/README.md
@@ -1,3 +1,0 @@
-# @radix-ng/primitives/tooltip-v2
-
-Secondary entry point of `@radix-ng/primitives`. It can be used by importing from `@radix-ng/primitives/tooltip-v2`.
