--- packages/primitives/tooltip-v2/ng-package.json
@@ -1,5 +1,0 @@
-{
-    "lib": {
-        "entryFile": "index.ts"
-    }
-}
