--- packages/primitives/tooltip-v2/src/tooltip-arrow.token.ts
@@ -1,4 +1,0 @@
-import { InjectionToken } from '@angular/core';
-import { RdxTooltipArrowDirective } from './tooltip-arrow.directive';
-
-export const RdxTooltipArrowToken = new InjectionToken<RdxTooltipArrowDirective>('RdxTooltipArrowToken');
