--- packages/primitives/tooltip-v2/src/tooltip.types.ts
@@ -1,21 +1,0 @@
-export enum RdxTooltipState {
-    OPEN = 'open',
-    CLOSED = 'closed'
-}
-
-export enum RdxTooltipAction {
-    OPEN = 'open',
-    CLOSE = 'close'
-}
-
-export enum RdxTooltipAttachDetachEvent {
-    ATTACH = 'attach',
-    DETACH = 'detach'
-}
-
-export enum RdxTooltipAnimationStatus {
-    OPEN_STARTED = 'open_started',
-    OPEN_ENDED = 'open_ended',
-    CLOSED_STARTED = 'closed_started',
-    CLOSED_ENDED = 'closed_ended'
-}
