--- packages/primitives/tooltip-v2/stories/tooltip-triggering.component.ts
@@ -1,200 +1,0 @@
-import { Component, signal, viewChild } from '@angular/core';
-import { FormsModule } from '@angular/forms';
-import { LucideAngularModule, MountainSnow, TriangleAlert, X } from 'lucide-angular';
-import { RdxTooltipModule } from '../index';
-import { RdxTooltipContentAttributesComponent } from '../src/tooltip-content-attributes.component';
-import { provideRdxCdkEventService } from '../src/utils/cdk-event.service';
-import { containerAlert } from './utils/constants';
-import { OptionPanelBase } from './utils/option-panel-base.class';
-import styles from './utils/styles.constants';
-import { WithOptionPanelComponent } from './utils/with-option-panel.component';
-
-@Component({
-    selector: 'rdx-tooltip-triggering',
-    providers: [provideRdxCdkEventService()],
-    imports: [
-        FormsModule,
-        RdxTooltipModule,
-        LucideAngularModule,
-        RdxTooltipContentAttributesComponent,
-        WithOptionPanelComponent
-    ],
-    styles: styles(),
-    template: `
-        <p class="ExampleSubtitle">Initially closed</p>
-        <tooltip-with-option-panel
-            [arrowWidth]="arrowWidth()"
-            [arrowHeight]="arrowHeight()"
-            [openDelay]="openDelay()"
-            [closeDelay]="closeDelay()"
-            (onOverlayEscapeKeyDownDisabledChange)="onOverlayEscapeKeyDownDisabled.set($event)"
-            (onOverlayOutsideClickDisabledChange)="onOverlayOutsideClickDisabled.set($event)"
-            (arrowWidthChange)="arrowWidth.set($event)"
-            (arrowHeightChange)="arrowHeight.set($event)"
-            (openDelayChange)="openDelay.set($event)"
-            (closeDelayChange)="closeDelay.set($event)"
-        >
-            <div class="ParamsContainer">
-                <button (mouseup)="triggerOpenFalse()" type="button">Open: {{ isOpenFalse() }}</button>
-                onOpenChange count: {{ counterOpenFalse() }}
-            </div>
-
-            <div class="ParamsContainer">
-                <input
-                    [ngModel]="externalControlFalse()"
-                    (ngModelChange)="externalControlFalse.set($event)"
-                    type="checkbox"
-                />
-                External control
-            </div>
-
-            <div class="ContainerAlerts">
-                <lucide-angular [img]="TriangleAlert" size="16" />
-                {{ containerAlert }}
-            </div>
-            <div class="container">
-                <ng-container
-                    #root1="rdxTooltipRoot"
-                    [open]="isOpenFalse()"
-                    [openDelay]="openDelay()"
-                    [closeDelay]="closeDelay()"
-                    [externalControl]="externalControlFalse()"
-                    rdxTooltipRoot
-                >
-                    <button class="reset IconButton" rdxTooltipTrigger>
-                        <lucide-angular [img]="MountainSnowIcon" size="16" style="display: flex" />
-                    </button>
-
-                    <ng-template
-                        [sideOffset]="8"
-                        [onOverlayEscapeKeyDownDisabled]="onOverlayEscapeKeyDownDisabled()"
-                        [onOverlayOutsideClickDisabled]="onOverlayOutsideClickDisabled()"
-                        (onOpen)="countOpenFalse(true)"
-                        (onClosed)="countOpenFalse(false)"
-                        rdxTooltipContent
-                    >
-                        <div class="TooltipContent" rdxTooltipContentAttributes>
-                            <button class="reset TooltipClose" rdxTooltipClose aria-label="Close">
-                                <lucide-angular [img]="XIcon" size="12" style="display: flex" />
-                            </button>
-                            Add to library
-                            <div
-                                class="TooltipArrow"
-                                [width]="arrowWidth()"
-                                [height]="arrowHeight()"
-                                rdxTooltipArrow
-                            ></div>
-                        </div>
-                    </ng-template>
-                </ng-container>
-            </div>
-            <div class="TooltipId">ID: {{ rootDirective1()?.uniqueId() }}</div>
-        </tooltip-with-option-panel>
-
-        <p class="ExampleSubtitle">Initially open</p>
-        <tooltip-with-option-panel
-            [arrowWidth]="arrowWidth()"
-            [arrowHeight]="arrowHeight()"
-            [openDelay]="openDelay()"
-            [closeDelay]="closeDelay()"
-            (onOverlayEscapeKeyDownDisabledChange)="onOverlayEscapeKeyDownDisabled.set($event)"
-            (onOverlayOutsideClickDisabledChange)="onOverlayOutsideClickDisabled.set($event)"
-            (arrowWidthChange)="arrowWidth.set($event)"
-            (arrowHeightChange)="arrowHeight.set($event)"
-            (openDelayChange)="openDelay.set($event)"
-            (closeDelayChange)="closeDelay.set($event)"
-        >
-            <div class="ParamsContainer">
-                <button (mouseup)="triggerOpenTrue()" type="button">Open: {{ isOpenTrue() }}</button>
-                <span>onOpenChange count: {{ counterOpenTrue() }}</span>
-            </div>
-
-            <div class="ParamsContainer">
-                <input
-                    [ngModel]="externalControlTrue()"
-                    (ngModelChange)="externalControlTrue.set($event)"
-                    type="checkbox"
-                />
-                External control
-            </div>
-
-            <div class="ContainerAlerts">
-                <lucide-angular [img]="TriangleAlert" size="16" />
-                {{ containerAlert }}
-            </div>
-            <div class="container">
-                <ng-container
-                    #root2="rdxTooltipRoot"
-                    [open]="isOpenTrue()"
-                    [openDelay]="openDelay()"
-                    [closeDelay]="closeDelay()"
-                    [externalControl]="externalControlTrue()"
-                    rdxTooltipRoot
-                >
-                    <button class="reset IconButton" rdxTooltipTrigger>
-                        <lucide-angular [img]="MountainSnowIcon" size="16" style="display: flex" />
-                    </button>
-
-                    <ng-template
-                        [sideOffset]="8"
-                        [onOverlayEscapeKeyDownDisabled]="onOverlayEscapeKeyDownDisabled()"
-                        [onOverlayOutsideClickDisabled]="onOverlayOutsideClickDisabled()"
-                        (onOpen)="countOpenTrue(true)"
-                        (onClosed)="countOpenTrue(false)"
-                        rdxTooltipContent
-                    >
-                        <div class="TooltipContent" rdxTooltipContentAttributes>
-                            <button class="reset TooltipClose" rdxTooltipClose aria-label="Close">
-                                <lucide-angular [img]="XIcon" size="12" style="display: flex" />
-                            </button>
-                            Add to library
-                            <div
-                                class="TooltipArrow"
-                                [width]="arrowWidth()"
-                                [height]="arrowHeight()"
-                                rdxTooltipArrow
-                            ></div>
-                        </div>
-                    </ng-template>
-                </ng-container>
-            </div>
-            <div class="TooltipId">ID: {{ rootDirective2()?.uniqueId() }}</div>
-        </tooltip-with-option-panel>
-    `
-})
-export class RdxTooltipTriggeringComponent extends OptionPanelBase {
-    readonly rootDirective1 = viewChild('root1');
-    readonly rootDirective2 = viewChild('root2');
-
-    readonly MountainSnowIcon = MountainSnow;
-    readonly XIcon = X;
-
-    isOpenFalse = signal(false);
-    counterOpenFalse = signal(0);
-    externalControlFalse = signal(true);
-
-    isOpenTrue = signal(true);
-    counterOpenTrue = signal(0);
-    externalControlTrue = signal(true);
-
-    triggerOpenFalse(): void {
-        this.isOpenFalse.update((value) => !value);
-    }
-
-    countOpenFalse(open: boolean): void {
-        this.isOpenFalse.set(open);
-        this.counterOpenFalse.update((value) => value + 1);
-    }
-
-    triggerOpenTrue(): void {
-        this.isOpenTrue.update((value) => !value);
-    }
-
-    countOpenTrue(open: boolean): void {
-        this.isOpenTrue.set(open);
-        this.counterOpenTrue.update((value) => value + 1);
-    }
-
-    protected readonly containerAlert = containerAlert;
-    protected readonly TriangleAlert = TriangleAlert;
-}
