+++ packages/primitives/tooltip/README.md
@@ -1,1 +1,3 @@
 # @radix-ng/primitives/tooltip
+
+Secondary entry point of `@radix-ng/primitives`. It can be used by importing from `@radix-ng/primitives/tooltip`.
