+++ packages/primitives/tooltip/index.ts
@@ -1,23 +1,28 @@
 import { NgModule } from '@angular/core';
+import { RdxTooltipAnchorDirective } from './src/tooltip-anchor.directive';
 import { RdxTooltipArrowDirective } from './src/tooltip-arrow.directive';
-import { RdxTooltipContentAttributesDirective } from './src/tooltip-content-attributes.directive';
+import { RdxTooltipCloseDirective } from './src/tooltip-close.directive';
+import { RdxTooltipContentAttributesComponent } from './src/tooltip-content-attributes.component';
 import { RdxTooltipContentDirective } from './src/tooltip-content.directive';
 import { RdxTooltipRootDirective } from './src/tooltip-root.directive';
 import { RdxTooltipTriggerDirective } from './src/tooltip-trigger.directive';
 
+export * from './src/tooltip-anchor.directive';
 export * from './src/tooltip-arrow.directive';
-export * from './src/tooltip-content-attributes.directive';
+export * from './src/tooltip-close.directive';
+export * from './src/tooltip-content-attributes.component';
 export * from './src/tooltip-content.directive';
 export * from './src/tooltip-root.directive';
 export * from './src/tooltip-trigger.directive';
-export * from './src/tooltip.types';
 
 const _imports = [
     RdxTooltipArrowDirective,
+    RdxTooltipCloseDirective,
     RdxTooltipContentDirective,
     RdxTooltipTriggerDirective,
-    RdxTooltipContentAttributesDirective,
-    RdxTooltipRootDirective
+    RdxTooltipRootDirective,
+    RdxTooltipAnchorDirective,
+    RdxTooltipContentAttributesComponent
 ];
 
 @NgModule({
