+++ packages/primitives/tooltip/src/tooltip-close.directive.ts
