+++ packages/primitives/tooltip/src/tooltip-content-attributes.component.ts
