--- packages/primitives/tooltip/src/tooltip-content.token.ts
@@ -1,4 +1,0 @@
-import { InjectionToken } from '@angular/core';
-import { RdxTooltipContentDirective } from './tooltip-content.directive';
-
-export const RdxTooltipContentToken = new InjectionToken<RdxTooltipContentDirective>('RdxTooltipContentToken');
