+++ packages/primitives/tooltip/src/tooltip-root.directive.ts
@@ -1,363 +1,421 @@
 import { BooleanInput, NumberInput } from '@angular/cdk/coercion';
-import { ConnectedPosition, Overlay, OverlayRef, PositionStrategy } from '@angular/cdk/overlay';
-import { TemplatePortal } from '@angular/cdk/portal';
-import { isPlatformBrowser } from '@angular/common';
 import {
+    afterNextRender,
     booleanAttribute,
     computed,
     contentChild,
     DestroyRef,
     Directive,
     effect,
-    forwardRef,
     inject,
-    InjectionToken,
     input,
     numberAttribute,
-    OnInit,
-    output,
-    PLATFORM_ID,
     signal,
     untracked,
-    ViewContainerRef,
-    ViewRef
+    ViewContainerRef
 } from '@angular/core';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
-import { injectDocument, injectWindow } from '@radix-ng/primitives/core';
-import { asyncScheduler, filter, take } from 'rxjs';
-import { RdxTooltipContentToken } from './tooltip-content.token';
+import { debounce, map, Subject, tap, timer } from 'rxjs';
+import { RdxTooltipAnchorDirective } from './tooltip-anchor.directive';
+import { RdxTooltipAnchorToken } from './tooltip-anchor.token';
+import { RdxTooltipArrowToken } from './tooltip-arrow.token';
+import { RdxTooltipCloseToken } from './tooltip-close.token';
+import { RdxTooltipContentAttributesToken } from './tooltip-content-attributes.token';
+import { RdxTooltipContentDirective } from './tooltip-content.directive';
 import { RdxTooltipTriggerDirective } from './tooltip-trigger.directive';
-import { injectTooltipConfig } from './tooltip.config';
-import { RdxTooltipState } from './tooltip.types';
+import {
+    RdxTooltipAction,
+    RdxTooltipAnimationStatus,
+    RdxTooltipAttachDetachEvent,
+    RdxTooltipState
+} from './tooltip.types';
+import { injectRdxCdkEventService } from './utils/cdk-event.service';
 
-export const RdxTooltipRootToken = new InjectionToken<RdxTooltipRootDirective>('RdxTooltipRootToken');
-
-export function injectTooltipRoot(): RdxTooltipRootDirective {
-    return inject(RdxTooltipRootToken);
-}
+let nextId = 0;
 
 @Directive({
     selector: '[rdxTooltipRoot]',
-    providers: [
-        {
-            provide: RdxTooltipRootToken,
-            useExisting: forwardRef(() => RdxTooltipRootDirective)
-        }
-    ],
     exportAs: 'rdxTooltipRoot'
 })
-export class RdxTooltipRootDirective implements OnInit {
+export class RdxTooltipRootDirective {
     /** @ignore */
-    private readonly viewContainerRef = inject(ViewContainerRef);
+    readonly uniqueId = signal(++nextId);
     /** @ignore */
-    private readonly destroyRef = inject(DestroyRef);
-    /** @ignore */
-    private readonly overlay = inject(Overlay);
-    /** @ignore */
-    private readonly platformId = inject(PLATFORM_ID);
-    /** @ignore */
-    private readonly document = injectDocument();
-    /** @ignore */
-    private readonly window = injectWindow();
-    /** @ignore */
-    readonly tooltipConfig = injectTooltipConfig();
+    readonly name = computed(() => `rdx-tooltip-root-${this.uniqueId()}`);
 
     /**
-     * The open state of the tooltip when it is initially rendered. Use when you do not need to control its open state.
+     * @description The anchor directive that comes form outside the tooltip rootDirective
+     * @default undefined
+     */
+    readonly anchor = input<RdxTooltipAnchorDirective | undefined>(void 0);
+    /**
+     * @description The open state of the tooltip when it is initially rendered. Use when you do not need to control its open state.
+     * @default false
      */
     readonly defaultOpen = input<boolean, BooleanInput>(false, { transform: booleanAttribute });
-
     /**
-     * The controlled open state of the tooltip. Must be used in conjunction with onOpenChange.
+     * @description The controlled state of the tooltip. `open` input take precedence of `defaultOpen` input.
+     * @default undefined
      */
     readonly open = input<boolean | undefined, BooleanInput>(void 0, { transform: booleanAttribute });
-
     /**
-     * Override the duration given to the configuration to customise the open delay for a specific tooltip.
+     * To customise the open delay for a specific tooltip.
      */
-    readonly delayDuration = input<number, NumberInput>(this.tooltipConfig.delayDuration, {
+    readonly openDelay = input<number, NumberInput>(500, {
         transform: numberAttribute
     });
-
-    /** @ignore */
-    readonly disableHoverableContent = input<boolean, BooleanInput>(
-        this.tooltipConfig.disableHoverableContent ?? false,
-        { transform: booleanAttribute }
-    );
-
     /**
-     * Event handler called when the open state of the tooltip changes.
+     * To customise the close delay for a specific tooltip.
      */
-    readonly onOpenChange = output<boolean>();
-
-    /** @ignore */
-    readonly isOpen = signal<boolean>(this.defaultOpen());
-    /** @ignore */
-    readonly isOpenDelayed = signal<boolean>(true);
-    /** @ignore */
-    readonly wasOpenDelayed = signal<boolean>(false);
-    /** @ignore */
-    readonly state = computed<RdxTooltipState>(() => {
-        const currentIsOpen = this.isOpen();
-        const currentWasOpenDelayed = this.wasOpenDelayed();
-
-        if (currentIsOpen) {
-            return currentWasOpenDelayed ? 'delayed-open' : 'instant-open';
-        }
-
-        return 'closed';
+    readonly closeDelay = input<number, NumberInput>(200, {
+        transform: numberAttribute
     });
-    /** @ignore */
-    readonly tooltipContentDirective = contentChild.required(RdxTooltipContentToken);
-    /** @ignore */
-    readonly tooltipTriggerDirective = contentChild.required(RdxTooltipTriggerDirective);
+    /**
+     * @description Whether to control the state of the tooltip from external. Use in conjunction with `open` input.
+     * @default undefined
+     */
+    readonly externalControl = input<boolean | undefined, BooleanInput>(void 0, { transform: booleanAttribute });
+    /**
+     * @description Whether to take into account CSS opening/closing animations.
+     * @default false
+     */
+    readonly cssAnimation = input<boolean, BooleanInput>(false, { transform: booleanAttribute });
+    /**
+     * @description Whether to take into account CSS opening animations. `cssAnimation` input must be set to 'true'
+     * @default false
+     */
+    readonly cssOpeningAnimation = input<boolean, BooleanInput>(false, { transform: booleanAttribute });
+    /**
+     * @description Whether to take into account CSS closing animations. `cssAnimation` input must be set to 'true'
+     * @default false
+     */
+    readonly cssClosingAnimation = input<boolean, BooleanInput>(false, { transform: booleanAttribute });
 
     /** @ignore */
-    private openTimer = 0;
-    /** @ignore */
-    private skipDelayTimer = 0;
-    /** @ignore */
-    private overlayRef?: OverlayRef;
-    /** @ignore */
-    private instance?: ViewRef;
-    /** @ignore */
-    private portal: TemplatePortal<unknown>;
-    /** @ignore */
-    private isControlledExternally = false;
+    readonly cssAnimationStatus = signal<RdxTooltipAnimationStatus | null>(null);
 
     /** @ignore */
-    ngOnInit(): void {
-        if (this.defaultOpen()) {
-            this.handleOpen();
-        }
+    readonly contentDirective = contentChild.required(RdxTooltipContentDirective);
+    /** @ignore */
+    readonly triggerDirective = contentChild.required(RdxTooltipTriggerDirective);
+    /** @ignore */
+    readonly arrowDirective = contentChild(RdxTooltipArrowToken);
+    /** @ignore */
+    readonly closeDirective = contentChild(RdxTooltipCloseToken);
+    /** @ignore */
+    readonly contentAttributesComponent = contentChild(RdxTooltipContentAttributesToken);
+    /** @ignore */
+    private readonly internalAnchorDirective = contentChild(RdxTooltipAnchorToken);
 
-        this.isControlledExternally = this.open() !== undefined;
+    /** @ignore */
+    readonly viewContainerRef = inject(ViewContainerRef);
+    /** @ignore */
+    readonly rdxCdkEventService = injectRdxCdkEventService();
+    /** @ignore */
+    readonly destroyRef = inject(DestroyRef);
+
+    /** @ignore */
+    readonly state = signal(RdxTooltipState.CLOSED);
+
+    /** @ignore */
+    readonly attachDetachEvent = signal(RdxTooltipAttachDetachEvent.DETACH);
+
+    /** @ignore */
+    private readonly isFirstDefaultOpen = signal(false);
+
+    /** @ignore */
+    readonly anchorDirective = computed(() => this.internalAnchorDirective() ?? this.anchor());
+
+    /** @ignore */
+    readonly actionSubject$ = new Subject<RdxTooltipAction>();
+
+    constructor() {
+        this.rdxCdkEventService?.registerPrimitive(this);
+        this.destroyRef.onDestroy(() => this.rdxCdkEventService?.deregisterPrimitive(this));
+        this.actionSubscription();
+        this.onStateChangeEffect();
+        this.onCssAnimationStatusChangeChangeEffect();
+        this.onOpenChangeEffect();
+        this.onIsFirstDefaultOpenChangeEffect();
+        this.onAnchorChangeEffect();
+        this.emitOpenOrClosedEventEffect();
+        afterNextRender({
+            write: () => {
+                if (this.defaultOpen() && !this.open()) {
+                    this.isFirstDefaultOpen.set(true);
+                }
+            }
+        });
     }
 
     /** @ignore */
-    onTriggerEnter(): void {
-        if (this.isControlledExternally) {
-            return;
-        }
-
-        if (this.isOpenDelayed()) {
-            this.handleDelayedOpen();
-        } else {
-            this.handleOpen();
-        }
+    getAnimationParamsSnapshot() {
+        return {
+            cssAnimation: this.cssAnimation(),
+            cssOpeningAnimation: this.cssOpeningAnimation(),
+            cssClosingAnimation: this.cssClosingAnimation(),
+            cssAnimationStatus: this.cssAnimationStatus(),
+            attachDetachEvent: this.attachDetachEvent(),
+            state: this.state(),
+            canEmitOnOpenOrOnClosed: this.canEmitOnOpenOrOnClosed()
+        };
     }
 
     /** @ignore */
-    onTriggerLeave(): void {
-        this.clearTimeout(this.openTimer);
-        this.handleClose();
+    controlledExternally() {
+        return this.externalControl;
     }
 
     /** @ignore */
-    onOpen(): void {
-        this.clearTimeout(this.skipDelayTimer);
-        this.isOpenDelayed.set(false);
-    }
-
-    /** @ignore */
-    onClose(): void {
-        this.clearTimeout(this.skipDelayTimer);
-
-        if (isPlatformBrowser(this.platformId)) {
-            this.skipDelayTimer = this.window.setTimeout(() => {
-                this.isOpenDelayed.set(true);
-            }, this.tooltipConfig.skipDelayDuration);
-        }
+    firstDefaultOpen() {
+        return this.isFirstDefaultOpen();
     }
 
     /** @ignore */
     handleOpen(): void {
-        if (this.isControlledExternally) {
+        if (this.externalControl()) {
             return;
         }
-
-        this.wasOpenDelayed.set(false);
-        this.setOpen(true);
+        this.actionSubject$.next(RdxTooltipAction.OPEN);
     }
 
     /** @ignore */
-    handleClose(): void {
-        if (this.isControlledExternally) {
+    handleClose(closeButton?: boolean): void {
+        if (this.isFirstDefaultOpen()) {
+            this.isFirstDefaultOpen.set(false);
+        }
+        if (!closeButton && this.externalControl()) {
             return;
         }
-
-        this.clearTimeout(this.openTimer);
-        this.setOpen(false);
+        this.actionSubject$.next(RdxTooltipAction.CLOSE);
     }
 
     /** @ignore */
-    private handleOverlayKeydown(): void {
-        if (!this.overlayRef) {
+    handleToggle(): void {
+        if (this.externalControl()) {
             return;
         }
+        this.isOpen() ? this.handleClose() : this.handleOpen();
+    }
 
-        this.overlayRef
-            .keydownEvents()
-            .pipe(
-                filter((event) => event.key === 'Escape'),
-                takeUntilDestroyed(this.destroyRef)
-            )
-            .subscribe((event) => {
-                this.tooltipContentDirective().onEscapeKeyDown.emit(event);
+    /** @ignore */
+    isOpen(state?: RdxTooltipState) {
+        return (state ?? this.state()) === RdxTooltipState.OPEN;
+    }
 
-                if (!event.defaultPrevented) {
-                    this.handleClose();
+    /** @ignore */
+    private setState(state = RdxTooltipState.CLOSED): void {
+        if (state === this.state()) {
+            return;
+        }
+        this.state.set(state);
+    }
+
+    /** @ignore */
+    private openContent(): void {
+        this.contentDirective().open();
+        if (!this.cssAnimation() || !this.cssOpeningAnimation()) {
+            this.cssAnimationStatus.set(null);
+        }
+    }
+
+    /** @ignore */
+    private closeContent(): void {
+        this.contentDirective().close();
+        if (!this.cssAnimation() || !this.cssClosingAnimation()) {
+            this.cssAnimationStatus.set(null);
+        }
+    }
+
+    /** @ignore */
+    private emitOnOpen(): void {
+        this.contentDirective().onOpen.emit();
+    }
+
+    /** @ignore */
+    private emitOnClosed(): void {
+        this.contentDirective().onClosed.emit();
+    }
+
+    /** @ignore */
+    private ifOpenOrCloseWithoutAnimations(state: RdxTooltipState) {
+        return (
+            !this.contentAttributesComponent() ||
+            !this.cssAnimation() ||
+            (this.cssAnimation() && !this.cssClosingAnimation() && state === RdxTooltipState.CLOSED) ||
+            (this.cssAnimation() && !this.cssOpeningAnimation() && state === RdxTooltipState.OPEN) ||
+            // !this.cssAnimationStatus() ||
+            (this.cssOpeningAnimation() &&
+                state === RdxTooltipState.OPEN &&
+                [RdxTooltipAnimationStatus.OPEN_STARTED].includes(this.cssAnimationStatus()!)) ||
+            (this.cssClosingAnimation() &&
+                state === RdxTooltipState.CLOSED &&
+                [RdxTooltipAnimationStatus.CLOSED_STARTED].includes(this.cssAnimationStatus()!))
+        );
+    }
+
+    /** @ignore */
+    private ifOpenOrCloseWithAnimations(cssAnimationStatus: RdxTooltipAnimationStatus | null) {
+        return (
+            this.contentAttributesComponent() &&
+            this.cssAnimation() &&
+            cssAnimationStatus &&
+            ((this.cssOpeningAnimation() &&
+                this.state() === RdxTooltipState.OPEN &&
+                [RdxTooltipAnimationStatus.OPEN_ENDED].includes(cssAnimationStatus)) ||
+                (this.cssClosingAnimation() &&
+                    this.state() === RdxTooltipState.CLOSED &&
+                    [RdxTooltipAnimationStatus.CLOSED_ENDED].includes(cssAnimationStatus)))
+        );
+    }
+
+    /** @ignore */
+    private openOrClose(state: RdxTooltipState) {
+        const isOpen = this.isOpen(state);
+        isOpen ? this.openContent() : this.closeContent();
+    }
+
+    /** @ignore */
+    private emitOnOpenOrOnClosed(state: RdxTooltipState) {
+        this.isOpen(state)
+            ? this.attachDetachEvent() === RdxTooltipAttachDetachEvent.ATTACH && this.emitOnOpen()
+            : this.attachDetachEvent() === RdxTooltipAttachDetachEvent.DETACH && this.emitOnClosed();
+    }
+
+    /** @ignore */
+    private canEmitOnOpenOrOnClosed() {
+        return (
+            !this.cssAnimation() ||
+            (!this.cssOpeningAnimation() && this.state() === RdxTooltipState.OPEN) ||
+            (this.cssOpeningAnimation() &&
+                this.state() === RdxTooltipState.OPEN &&
+                this.cssAnimationStatus() === RdxTooltipAnimationStatus.OPEN_ENDED) ||
+            (!this.cssClosingAnimation() && this.state() === RdxTooltipState.CLOSED) ||
+            (this.cssClosingAnimation() &&
+                this.state() === RdxTooltipState.CLOSED &&
+                this.cssAnimationStatus() === RdxTooltipAnimationStatus.CLOSED_ENDED)
+        );
+    }
+
+    /** @ignore */
+    private onStateChangeEffect() {
+        let isFirst = true;
+        effect(() => {
+            const state = this.state();
+            untracked(() => {
+                if (isFirst) {
+                    isFirst = false;
+                    return;
+                }
+                if (!this.ifOpenOrCloseWithoutAnimations(state)) {
+                    return;
+                }
+                this.openOrClose(state);
+            });
+        }, {});
+    }
+
+    /** @ignore */
+    private onCssAnimationStatusChangeChangeEffect() {
+        let isFirst = true;
+        effect(() => {
+            const cssAnimationStatus = this.cssAnimationStatus();
+            untracked(() => {
+                if (isFirst) {
+                    isFirst = false;
+                    return;
+                }
+                if (!this.ifOpenOrCloseWithAnimations(cssAnimationStatus)) {
+                    return;
+                }
+                this.openOrClose(this.state());
+            });
+        });
+    }
+
+    /** @ignore */
+    private emitOpenOrClosedEventEffect() {
+        let isFirst = true;
+        effect(() => {
+            this.attachDetachEvent();
+            this.cssAnimationStatus();
+            untracked(() => {
+                if (isFirst) {
+                    isFirst = false;
+                    return;
+                }
+                const canEmitOpenClose = untracked(() => this.canEmitOnOpenOrOnClosed());
+                if (!canEmitOpenClose) {
+                    return;
+                }
+                this.emitOnOpenOrOnClosed(this.state());
+            });
+        });
+    }
+
+    /** @ignore */
+    private onOpenChangeEffect() {
+        effect(() => {
+            const open = this.open();
+            untracked(() => {
+                this.setState(open ? RdxTooltipState.OPEN : RdxTooltipState.CLOSED);
+            });
+        });
+    }
+
+    /** @ignore */
+    private onIsFirstDefaultOpenChangeEffect() {
+        const effectRef = effect(() => {
+            const defaultOpen = this.defaultOpen();
+            untracked(() => {
+                if (!defaultOpen || this.open()) {
+                    effectRef.destroy();
+                    return;
+                }
+                this.handleOpen();
+            });
+        });
+    }
+
+    /** @ignore */
+    private onAnchorChangeEffect = () => {
+        effect(() => {
+            const anchor = this.anchor();
+            untracked(() => {
+                if (anchor) {
+                    anchor.setRoot(this);
                 }
             });
-    }
-
-    /** @ignore */
-    private handlePointerDownOutside(): void {
-        if (!this.overlayRef) {
-            return;
-        }
-
-        this.overlayRef
-            .outsidePointerEvents()
-            .pipe(takeUntilDestroyed(this.destroyRef))
-            .subscribe((event) => this.tooltipContentDirective().onPointerDownOutside.emit(event));
-    }
-
-    /** @ignore */
-    private handleDelayedOpen(): void {
-        this.clearTimeout(this.openTimer);
-
-        if (isPlatformBrowser(this.platformId)) {
-            this.openTimer = this.window.setTimeout(() => {
-                this.wasOpenDelayed.set(true);
-                this.setOpen(true);
-            }, this.delayDuration());
-        }
-    }
-
-    /** @ignore */
-    private setOpen(open = false): void {
-        if (open) {
-            this.onOpen();
-
-            this.document.dispatchEvent(new CustomEvent('tooltip.open'));
-        } else {
-            this.onClose();
-        }
-
-        this.isOpen.set(open);
-        this.onOpenChange.emit(open);
-    }
-
-    /** @ignore */
-    private createOverlayRef(): OverlayRef {
-        if (this.overlayRef) {
-            return this.overlayRef;
-        }
-
-        this.overlayRef = this.overlay.create({
-            direction: undefined,
-            positionStrategy: this.getPositionStrategy(this.tooltipContentDirective().position()),
-            scrollStrategy: this.overlay.scrollStrategies.close()
         });
+    };
 
-        this.overlayRef
-            .detachments()
-            .pipe(take(1), takeUntilDestroyed(this.destroyRef))
-            .subscribe(() => this.detach());
-
-        this.handleOverlayKeydown();
-        this.handlePointerDownOutside();
-
-        return this.overlayRef;
+    /** @ignore */
+    private actionSubscription() {
+        this.actionSubject$
+            .asObservable()
+            .pipe(
+                map((action) => {
+                    console.log(action);
+                    switch (action) {
+                        case RdxTooltipAction.OPEN:
+                            return { action, duration: this.openDelay() };
+                        case RdxTooltipAction.CLOSE:
+                            return { action, duration: this.closeDelay() };
+                    }
+                }),
+                debounce((config) => timer(config.duration)),
+                tap((config) => {
+                    switch (config.action) {
+                        case RdxTooltipAction.OPEN:
+                            this.setState(RdxTooltipState.OPEN);
+                            break;
+                        case RdxTooltipAction.CLOSE:
+                            this.setState(RdxTooltipState.CLOSED);
+                            break;
+                    }
+                }),
+                takeUntilDestroyed()
+            )
+            .subscribe();
     }
-
-    /** @ignore */
-    private show(): void {
-        this.overlayRef = this.createOverlayRef();
-
-        this.detach();
-
-        this.portal =
-            this.portal ||
-            new TemplatePortal(this.tooltipContentDirective().templateRef, this.viewContainerRef, {
-                state: this.state,
-                side: this.tooltipContentDirective().side
-            });
-
-        this.instance = this.overlayRef.attach(this.portal);
-    }
-
-    /** @ignore */
-    private detach(): void {
-        if (this.overlayRef?.hasAttached()) {
-            this.overlayRef.detach();
-        }
-    }
-
-    /** @ignore */
-    private hide(): void {
-        if (this.isControlledExternally && this.open()) {
-            return;
-        }
-
-        asyncScheduler.schedule(() => {
-            this.instance?.destroy();
-        }, this.tooltipConfig.hideDelayDuration ?? 0);
-    }
-
-    /** @ignore */
-    private getPositionStrategy(connectedPosition: ConnectedPosition): PositionStrategy {
-        return this.overlay
-            .position()
-            .flexibleConnectedTo(this.tooltipTriggerDirective().elementRef)
-            .withFlexibleDimensions(false)
-            .withPositions([
-                connectedPosition
-            ])
-            .withLockedPosition();
-    }
-
-    /** @ignore */
-    private clearTimeout(timeoutId: number): void {
-        if (isPlatformBrowser(this.platformId)) {
-            this.window.clearTimeout(timeoutId);
-        }
-    }
-
-    /** @ignore */
-    private readonly onIsOpenChangeEffect = effect(() => {
-        const isOpen = this.isOpen();
-
-        untracked(() => {
-            if (isOpen) {
-                this.show();
-            } else {
-                this.hide();
-            }
-        });
-    });
-
-    /** @ignore */
-    private readonly onPositionChangeEffect = effect(() => {
-        const position = this.tooltipContentDirective().position();
-
-        if (this.overlayRef) {
-            const positionStrategy = this.getPositionStrategy(position);
-
-            this.overlayRef.updatePositionStrategy(positionStrategy);
-        }
-    });
-
-    /** @ignore */
-    private readonly onOpenChangeEffect = effect(() => {
-        const currentOpen = this.open();
-        this.isControlledExternally = currentOpen !== undefined;
-
-        untracked(() => {
-            if (this.isControlledExternally) {
-                this.setOpen(currentOpen);
-            }
-        });
-    });
 }
