+++ packages/primitives/tooltip/stories/tooltip-initially-open.component.ts
