+++ packages/primitives/tooltip/stories/tooltip-positioning.component.ts
@@ -1,177 +1,122 @@
-import { Component } from '@angular/core';
+import { Component, signal, viewChild } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { RdxPositionAlign, RdxPositionSide } from '@radix-ng/primitives/core';
-import { LucideAngularModule, Plus } from 'lucide-angular';
-import { RdxTooltipModule } from '../index';
+import { LucideAngularModule, MountainSnow, TriangleAlert, X } from 'lucide-angular';
+import { RdxTooltipModule, RdxTooltipRootDirective } from '../index';
+import { RdxTooltipContentAttributesComponent } from '../src/tooltip-content-attributes.component';
+import { provideRdxCdkEventService } from '../src/utils/cdk-event.service';
+import { containerAlert } from './utils/constants';
+import { OptionPanelBase } from './utils/option-panel-base.class';
+import styles from './utils/styles.constants';
+import { WithOptionPanelComponent } from './utils/with-option-panel.component';
 
 @Component({
     selector: 'rdx-tooltip-positioning',
+    providers: [provideRdxCdkEventService()],
     imports: [
         FormsModule,
         RdxTooltipModule,
-        LucideAngularModule
+        LucideAngularModule,
+        RdxTooltipContentAttributesComponent,
+        WithOptionPanelComponent
     ],
-    styles: `
-        .container {
-            height: 150px;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-        }
-
-        /* reset */
-        button {
-            all: unset;
-        }
-
-        .TooltipContent {
-            border-radius: 4px;
-            padding: 10px 15px;
-            font-size: 15px;
-            line-height: 1;
-            color: var(--violet-11);
-            background-color: white;
-            box-shadow:
-                hsl(206 22% 7% / 35%) 0px 10px 38px -10px,
-                hsl(206 22% 7% / 20%) 0px 10px 20px -15px;
-            user-select: none;
-            animation-duration: 400ms;
-            animation-timing-function: cubic-bezier(0.16, 1, 0.3, 1);
-            will-change: transform, opacity;
-        }
-        .TooltipContent[data-state='delayed-open'][data-side='top'] {
-            animation-name: slideDownAndFade;
-        }
-        .TooltipContent[data-state='delayed-open'][data-side='right'] {
-            animation-name: slideLeftAndFade;
-        }
-        .TooltipContent[data-state='delayed-open'][data-side='bottom'] {
-            animation-name: slideUpAndFade;
-        }
-        .TooltipContent[data-state='delayed-open'][data-side='left'] {
-            animation-name: slideRightAndFade;
-        }
-
-        .TooltipArrow {
-            fill: white;
-        }
-
-        .IconButton {
-            font-family: inherit;
-            border-radius: 100%;
-            height: 35px;
-            width: 35px;
-            display: inline-flex;
-            align-items: center;
-            justify-content: center;
-            color: var(--violet-11);
-            background-color: white;
-            box-shadow: 0 2px 10px var(--black-a7);
-        }
-        .IconButton:hover {
-            background-color: var(--violet-3);
-        }
-        .IconButton:focus {
-            box-shadow: 0 0 0 2px black;
-        }
-
-        @keyframes slideUpAndFade {
-            from {
-                opacity: 0;
-                transform: translateY(2px);
-            }
-            to {
-                opacity: 1;
-                transform: translateY(0);
-            }
-        }
-
-        @keyframes slideRightAndFade {
-            from {
-                opacity: 0;
-                transform: translateX(-2px);
-            }
-            to {
-                opacity: 1;
-                transform: translateX(0);
-            }
-        }
-
-        @keyframes slideDownAndFade {
-            from {
-                opacity: 0;
-                transform: translateY(-2px);
-            }
-            to {
-                opacity: 1;
-                transform: translateY(0);
-            }
-        }
-
-        @keyframes slideLeftAndFade {
-            from {
-                opacity: 0;
-                transform: translateX(2px);
-            }
-            to {
-                opacity: 1;
-                transform: translateX(0);
-            }
-        }
-
-        /* =============== Params layout =============== */
-
-        .ParamsContainer {
-            display: flex;
-            column-gap: 8px;
-            color: var(--white-a12);
-            margin-bottom: 32px;
-        }
-    `,
+    styles: styles(),
     template: `
-        <div class="ParamsContainer">
-            Side:
-            <select [(ngModel)]="selectedSide">
-                <option [value]="sides.Top">{{ sides.Top }}</option>
-                <option [value]="sides.Bottom">{{ sides.Bottom }}</option>
-                <option [value]="sides.Left">{{ sides.Left }}</option>
-                <option [value]="sides.Right">{{ sides.Right }}</option>
-            </select>
-            Align:
-            <select [(ngModel)]="selectedAlign">
-                <option [value]="aligns.Center">{{ aligns.Center }}</option>
-                <option [value]="aligns.Start">{{ aligns.Start }}</option>
-                <option [value]="aligns.End">{{ aligns.End }}</option>
-            </select>
-            SideOffset:
-            <input [(ngModel)]="sideOffset" type="number" />
-        </div>
+        <tooltip-with-option-panel
+            [arrowWidth]="arrowWidth()"
+            [arrowHeight]="arrowHeight()"
+            [openDelay]="openDelay()"
+            [closeDelay]="closeDelay()"
+            (onOverlayEscapeKeyDownDisabledChange)="onOverlayEscapeKeyDownDisabled.set($event)"
+            (onOverlayOutsideClickDisabledChange)="onOverlayOutsideClickDisabled.set($event)"
+            (arrowWidthChange)="arrowWidth.set($event)"
+            (arrowHeightChange)="arrowHeight.set($event)"
+            (openDelayChange)="openDelay.set($event)"
+            (closeDelayChange)="closeDelay.set($event)"
+        >
+            <div class="ParamsContainer">
+                Side:
+                <select [ngModel]="selectedSide()" (ngModelChange)="selectedSide.set($event)">
+                    <option [value]="sides.Top">{{ sides.Top }}</option>
+                    <option [value]="sides.Bottom">{{ sides.Bottom }}</option>
+                    <option [value]="sides.Left">{{ sides.Left }}</option>
+                    <option [value]="sides.Right">{{ sides.Right }}</option>
+                </select>
+                Align:
+                <select [ngModel]="selectedAlign()" (ngModelChange)="selectedAlign.set($event)">
+                    <option [value]="aligns.Center">{{ aligns.Center }}</option>
+                    <option [value]="aligns.Start">{{ aligns.Start }}</option>
+                    <option [value]="aligns.End">{{ aligns.End }}</option>
+                </select>
+                SideOffset:
+                <input [ngModel]="sideOffset()" (ngModelChange)="sideOffset.set($event)" type="number" />
+                AlignOffset:
+                <input [ngModel]="alignOffset()" (ngModelChange)="alignOffset.set($event)" type="number" />
+            </div>
 
-        <div class="container">
-            <ng-container rdxTooltipRoot>
-                <button class="IconButton" #triggerElement rdxTooltipTrigger>
-                    <lucide-angular [img]="PlusIcon" size="16" style="display: flex" />
-                </button>
+            <div class="ParamsContainer">
+                <input
+                    [ngModel]="disableAlternatePositions()"
+                    (ngModelChange)="disableAlternatePositions.set($event)"
+                    type="checkbox"
+                />
+                Disable alternate positions (to see the result, scroll the page to make the tooltip cross the viewport
+                boundary)
+            </div>
 
-                <ng-template [sideOffset]="sideOffset" [side]="selectedSide" [align]="selectedAlign" rdxTooltipContent>
-                    <div class="TooltipContent" rdxTooltipContentAttributes>
-                        Add to library
-                        <br />
-                        or do nothing
-                        <div class="TooltipArrow" rdxTooltipArrow></div>
-                    </div>
-                </ng-template>
-            </ng-container>
-        </div>
+            <div class="ContainerAlerts">
+                <lucide-angular [img]="TriangleAlert" size="16" />
+                {{ containerAlert }}
+            </div>
+            <div class="container">
+                <ng-container [openDelay]="openDelay()" [closeDelay]="closeDelay()" rdxTooltipRoot>
+                    <button class="reset IconButton" rdxTooltipTrigger>
+                        <lucide-angular [img]="MountainSnowIcon" size="16" style="display: flex" />
+                    </button>
+
+                    <ng-template
+                        [sideOffset]="sideOffset()"
+                        [alignOffset]="alignOffset()"
+                        [side]="selectedSide()"
+                        [align]="selectedAlign()"
+                        [alternatePositionsDisabled]="disableAlternatePositions()"
+                        [onOverlayEscapeKeyDownDisabled]="onOverlayEscapeKeyDownDisabled()"
+                        [onOverlayOutsideClickDisabled]="onOverlayOutsideClickDisabled()"
+                        rdxTooltipContent
+                    >
+                        <div class="TooltipContent" rdxTooltipContentAttributes>
+                            <button class="reset TooltipClose" rdxTooltipClose aria-label="Close">
+                                <lucide-angular [img]="XIcon" size="12" style="display: flex" />
+                            </button>
+                            Add to library
+                            <div
+                                class="TooltipArrow"
+                                [width]="arrowWidth()"
+                                [height]="arrowHeight()"
+                                rdxTooltipArrow
+                            ></div>
+                        </div>
+                    </ng-template>
+                </ng-container>
+            </div>
+            <div class="TooltipId">ID: {{ rootDirective()?.uniqueId() }}</div>
+        </tooltip-with-option-panel>
     `
 })
-export class RdxTooltipPositioningComponent {
-    readonly PlusIcon = Plus;
+export class RdxTooltipPositioningComponent extends OptionPanelBase {
+    readonly rootDirective = viewChild(RdxTooltipRootDirective);
 
-    selectedSide = RdxPositionSide.Top;
-    selectedAlign = RdxPositionAlign.Center;
-    sideOffset = 8;
+    readonly selectedSide = signal(RdxPositionSide.Top);
+    readonly selectedAlign = signal(RdxPositionAlign.Center);
+    readonly sideOffset = signal<number | undefined>(void 0);
+    readonly alignOffset = signal<number | undefined>(void 0);
+    readonly disableAlternatePositions = signal(false);
 
     readonly sides = RdxPositionSide;
     readonly aligns = RdxPositionAlign;
+
+    readonly MountainSnowIcon = MountainSnow;
+    readonly XIcon = X;
+    protected readonly containerAlert = containerAlert;
+    protected readonly TriangleAlert = TriangleAlert;
 }
