+++ packages/primitives/tooltip/stories/tooltip.stories.ts
@@ -1,8 +1,13 @@
 import { provideAnimations } from '@angular/platform-browser/animations';
 import { componentWrapperDecorator, Meta, moduleMetadata, StoryObj } from '@storybook/angular';
-import { LucideAngularModule, Plus } from 'lucide-angular';
+import { LucideAngularModule, MountainSnow, X } from 'lucide-angular';
 import { RdxTooltipModule } from '../index';
+import { RdxTooltipAnchorComponent } from './tooltip-anchor.component';
+import { RdxTooltipAnimationsComponent } from './tooltip-animations.component';
+import { RdxTooltipDefaultComponent } from './tooltip-default.component';
 import { RdxTooltipEventsComponent } from './tooltip-events.components';
+import { RdxTooltipInitiallyOpenComponent } from './tooltip-initially-open.component';
+import { RdxTooltipMultipleComponent } from './tooltip-multiple.component';
 import { RdxTooltipPositioningComponent } from './tooltip-positioning.component';
 import { RdxTooltipTriggeringComponent } from './tooltip-triggering.component';
 
@@ -14,11 +19,16 @@ export default {
         moduleMetadata({
             imports: [
                 RdxTooltipModule,
+                RdxTooltipDefaultComponent,
                 RdxTooltipEventsComponent,
                 RdxTooltipPositioningComponent,
                 RdxTooltipTriggeringComponent,
+                RdxTooltipMultipleComponent,
+                RdxTooltipAnimationsComponent,
+                RdxTooltipInitiallyOpenComponent,
+                RdxTooltipAnchorComponent,
                 LucideAngularModule,
-                LucideAngularModule.pick({ Plus })
+                LucideAngularModule.pick({ MountainSnow, X })
             ],
             providers: [provideAnimations()]
         }),
@@ -32,115 +42,6 @@ export default {
                 >
                     ${story}
                 </div>
-
-                <style>
-                    .container {
-                        height: 150px;
-                        display: flex;
-                        justify-content: center;
-                        align-items: center;
-                    }
-
-                    /* reset */
-                    button {
-                        all: unset;
-                    }
-
-                    .TooltipContent {
-                        border-radius: 4px;
-                        padding: 10px 15px;
-                        font-size: 15px;
-                        line-height: 1;
-                        color: var(--violet-11);
-                        background-color: white;
-                        box-shadow:
-                            hsl(206 22% 7% / 35%) 0px 10px 38px -10px,
-                            hsl(206 22% 7% / 20%) 0px 10px 20px -15px;
-                        user-select: none;
-                        animation-duration: 400ms;
-                        animation-timing-function: cubic-bezier(0.16, 1, 0.3, 1);
-                        will-change: transform, opacity;
-                    }
-                    .TooltipContent[data-state='delayed-open'][data-side='top'] {
-                        animation-name: slideDownAndFade;
-                    }
-                    .TooltipContent[data-state='delayed-open'][data-side='right'] {
-                        animation-name: slideLeftAndFade;
-                    }
-                    .TooltipContent[data-state='delayed-open'][data-side='bottom'] {
-                        animation-name: slideUpAndFade;
-                    }
-                    .TooltipContent[data-state='delayed-open'][data-side='left'] {
-                        animation-name: slideRightAndFade;
-                    }
-
-                    .TooltipArrow {
-                        fill: white;
-                    }
-
-                    .IconButton {
-                        font-family: inherit;
-                        border-radius: 100%;
-                        height: 35px;
-                        width: 35px;
-                        display: inline-flex;
-                        align-items: center;
-                        justify-content: center;
-                        color: var(--violet-11);
-                        background-color: white;
-                        box-shadow: 0 2px 10px var(--black-a7);
-                    }
-                    .IconButton:hover {
-                        background-color: var(--violet-3);
-                    }
-                    .IconButton:focus {
-                        box-shadow: 0 0 0 2px black;
-                    }
-
-                    @keyframes slideUpAndFade {
-                        from {
-                            opacity: 0;
-                            transform: translateY(2px);
-                        }
-                        to {
-                            opacity: 1;
-                            transform: translateY(0);
-                        }
-                    }
-
-                    @keyframes slideRightAndFade {
-                        from {
-                            opacity: 0;
-                            transform: translateX(-2px);
-                        }
-                        to {
-                            opacity: 1;
-                            transform: translateX(0);
-                        }
-                    }
-
-                    @keyframes slideDownAndFade {
-                        from {
-                            opacity: 0;
-                            transform: translateY(-2px);
-                        }
-                        to {
-                            opacity: 1;
-                            transform: translateY(0);
-                        }
-                    }
-
-                    @keyframes slideLeftAndFade {
-                        from {
-                            opacity: 0;
-                            transform: translateX(2px);
-                        }
-                        to {
-                            opacity: 1;
-                            transform: translateX(0);
-                        }
-                    }
-                </style>
             `
         )
     ]
@@ -151,20 +52,7 @@ type Story = StoryObj;
 export const Default: Story = {
     render: () => ({
         template: html`
-            <div class="container">
-                <ng-container rdxTooltipRoot>
-                    <button class="IconButton" rdxTooltipTrigger>
-                        <lucide-angular size="16" name="plus" style="display: flex;"></lucide-angular>
-                    </button>
-
-                    <ng-template rdxTooltipContent [sideOffset]="8">
-                        <div class="TooltipContent" rdxTooltipContentAttributes>
-                            Add to library
-                            <div class="TooltipArrow" rdxTooltipArrow></div>
-                        </div>
-                    </ng-template>
-                </ng-container>
-            </div>
+            <rdx-tooltip-default></rdx-tooltip-default>
         `
     })
 };
@@ -172,46 +60,7 @@ export const Default: Story = {
 export const Multiple: Story = {
     render: () => ({
         template: html`
-            <div class="container" style="gap: 40px;">
-                <ng-container rdxTooltipRoot>
-                    <button class="IconButton" rdxTooltipTrigger>
-                        <lucide-angular size="16" name="plus" style="display: flex;"></lucide-angular>
-                    </button>
-
-                    <ng-template rdxTooltipContent [sideOffset]="30">
-                        <div class="TooltipContent" rdxTooltipContentAttributes>
-                            Add to library
-                            <div class="TooltipArrow" rdxTooltipArrow></div>
-                        </div>
-                    </ng-template>
-                </ng-container>
-
-                <ng-container rdxTooltipRoot>
-                    <button class="IconButton" rdxTooltipTrigger>
-                        <lucide-angular size="16" name="plus" style="display: flex;"></lucide-angular>
-                    </button>
-
-                    <ng-template rdxTooltipContent [sideOffset]="8">
-                        <div class="TooltipContent" rdxTooltipContentAttributes>
-                            Add to library
-                            <div class="TooltipArrow" rdxTooltipArrow></div>
-                        </div>
-                    </ng-template>
-                </ng-container>
-
-                <ng-container rdxTooltipRoot>
-                    <button class="IconButton" rdxTooltipTrigger>
-                        <lucide-angular size="16" name="plus" style="display: flex;"></lucide-angular>
-                    </button>
-
-                    <ng-template rdxTooltipContent [sideOffset]="1">
-                        <div class="TooltipContent" rdxTooltipContentAttributes>
-                            Add to library
-                            <div class="TooltipArrow" rdxTooltipArrow></div>
-                        </div>
-                    </ng-template>
-                </ng-container>
-            </div>
+            <rdx-tooltip-multiple></rdx-tooltip-multiple>
         `
     })
 };
@@ -236,6 +85,30 @@ export const ExternalTriggering: Story = {
     render: () => ({
         template: html`
             <rdx-tooltip-triggering></rdx-tooltip-triggering>
+        `
+    })
+};
+
+export const Anchor: Story = {
+    render: () => ({
+        template: html`
+            <rdx-tooltip-anchor></rdx-tooltip-anchor>
+        `
+    })
+};
+
+export const InitiallyOpen: Story = {
+    render: () => ({
+        template: html`
+            <rdx-tooltip-initially-open></rdx-tooltip-initially-open>
+        `
+    })
+};
+
+export const Animations: Story = {
+    render: () => ({
+        template: html`
+            <rdx-tooltip-animations></rdx-tooltip-animations>
         `
     })
 };
